🚚 rename(/Answer): 정답(Lotto와 BonusLotto)을 담는 패키지명 Lotto -> Answer로 변경
