+++ src/test/java/lotto/model/RefereeTest.java
@@ -1,7 +1,7 @@
 package lotto.model;
 
-import lotto.model.Lotto.BonusLotto;
-import lotto.model.Lotto.Lotto;
+import lotto.model.Answer.BonusLotto;
+import lotto.model.Answer.Lotto;
 import org.junit.jupiter.api.*;
 
 import java.util.ArrayList;
